+++ package.json
@@ -27,7 +27,10 @@
     "next-auth": "^4.22.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "remark-html": "^15.0.2",
+    "remark-parse": "^10.0.2",
     "superjson": "^1.13.1",
+    "unified": "^11.0.3",
     "zod": "^3.21.4"
   },
   "devDependencies": {
