+++ src/components/StyledCheckbox.tsx
