+++ src/types/global.d.ts
@@ -1,0 +1,8 @@
+export class Question {
+  title = "";
+  body = "";
+  difficulty = 0;
+  category = "";
+}
+
+export type QuestionMap = Map<string, Question>;
